+++ avl_axi/_rmonitor.py
@@ -126,6 +126,8 @@ class ReadMonitor(avl.Monitor):
         :raises NotImplementedError: If the run phase is not implemented.
         """
 
+        await self.wait_on_reset()
+
         while True:
 
             tasks = []