+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+[v0.2.0] - 2025-10-12
+
+### Fixed
+ - [#2](https://github.com/projectapheleia/avl-axi/issues/2) Unique IDs and Tag ID create new dict instead of clearing class dict
+ - [#1](https://github.com/projectapheleia/avl-axi/issues/1) Modelsim does not support detecting parameters through VPI
+
 [v0.1.0] - 2025-09-28
 
 ### Added